+++ semcore/base-components/src/components/popper/createPopper.ts
@@ -13,8 +13,8 @@ import {
 import maxSize from 'popper-max-size-modifier';
 
 import arrowOffset from './arrowOffset';
+import { applyEndVariation } from './modifiers/applyEndVariation';
 import { applyMaxSize } from './modifiers/applyMaxSize';
-import { applyVariations } from './modifiers/applyVariations';
 import { cursorAnchoringModifier } from './modifiers/cursorAnchoring';
 
 const defaultModifiers = [
@@ -31,7 +31,7 @@ const defaultModifiers = [
   hide,
   maxSize,
   applyMaxSize,
-  applyVariations,
+  applyEndVariation,
 ];
 
 export default popperGenerator({ defaultModifiers });